requirements.txt
GitPython==2.1.8
python-dotenv==0.7.1